Framework.plot() using Graph.plot() to avoid code repetition
