+++ pyrigi/framework.py
@@ -271,124 +271,18 @@ class Framework(object):
         """
         return deepcopy(self._graph)
 
-    def resolve_edge_colors(
-        self, edge_color: Union(str, list[list[Edge]], dict[str : list[Edge]])
-    ) -> tuple[list, list]:
-        edge_list = self.graph().edge_list()
-        edge_list_ref = []
-        edge_color_array = []
-        colors = [
-            "red",
-            "green",
-            "blue",
-            "yellow",
-            "orange",
-            "purple",
-            "pink",
-            "lime",
-            "cyan",
-            "magenta",
-            "brown",
-            "darkblue",
-            "gold",
-            "lightgreen",
-            "violet",
-            "lightblue",
-            "orangered",
-            "olive",
-            "dodgerblue",
-        ]
-        color = ""
-        if isinstance(edge_color, str):
-            edge_list_ref = edge_list
-            for e in edge_list_ref:
-                edge_color_array.append(edge_color)
-        elif isinstance(edge_color, list):
-            edge_list_list = edge_color
-            for i in range(len(edge_list_list)):
-                if i >= len(colors):
-                    color = "black"
-                else:
-                    color = colors[i]
-                for e in edge_list_list[i]:
-                    if not self.graph().has_edge(e[0], e[1]):
-                        raise ValueError(
-                            "Input includes edge that is not part of the framework"
-                        )
-                    edge_color_array.append(color)
-                    edge_list_ref.append(e)
-            for e in edge_list:
-                if (e[0], e[1]) in edge_list_ref or (e[1], e[0]) in edge_list_ref:
-                    continue
-                else:
-                    edge_color_array.append("black")
-                    edge_list_ref.append(e)
-        elif isinstance(edge_color, dict):
-            color_edges_dict = edge_color
-            for color, edges in color_edges_dict.items():
-                for e in edges:
-                    if not self.graph().has_edge(e[0], e[1]):
-                        raise ValueError(
-                            "Input includes edge that is not part of the framework"
-                        )
-                    edge_color_array.append(color)
-                    edge_list_ref.append(e)
-            for e in edge_list:
-                if (e[0], e[1]) in edge_list_ref or (e[1], e[0]) in edge_list_ref:
-                    continue
-                else:
-                    edge_color_array.append("black")
-                    edge_list_ref.append(e)
-        return edge_color_array, edge_list_ref
-
     @doc_category("Plotting")
     def plot(
         self,
-        vertex_size: int = 300,
-        vertex_color: str = "#4584B6",
-        vertex_shape: str = "o",
-        vertex_labels: bool = True,
-        edge_width: float = 1.0,
-        edge_color: Union(str, list[list[Edge]], dict[str : list[Edge]]) = "black",
-        edge_style: str = "solid",
-        canvas_width: int = 6.4,
-        canvas_height: int = 4.8,
-        aspect_ratio: float = 1.0,
         **kwargs,
     ) -> None:
         """
         Plot the framework.
 
+        For the various formatting options, see :meth:`~Graph.plot`.
+
         Parameters
         ----------
-        vertex_size:
-            The size of the vertex. By default 300.
-        vertex_color:
-            The color of the vertex. Color can be string or rgb (or rgba)
-            tuple of floats from 0-1. By default '#4584B6'
-        vertex_shape:
-            The shape of the vertex. Specification is as matplotlib.scatter
-            marker, one of 'so^>v<dph8'. By default 'o'.
-        vertex_labels:
-            If True, vertex labels are displayed. By default True.
-        edge_width:
-            The width of the edge. By default 1.0.
-        edge_color:
-            The color of the edge. Color can be string or rgb (or rgba) tuple
-            of floats from 0-1. By default 'k'.
-        edge_style:
-            Edge line style e.g.: '-', '–', '-', ':' or words like 'solid' or
-            'dashed'. By default '-'.
-        font_size:
-            The size of the font used for the labels. By default 12.
-        font_color:
-            The color of the font used for the labels. By default 'k'.
-        canvas_width:
-            The width of the canvas in inches. By default 6.4.
-        canvas_height:
-            The height of the canvas in inches. By default 4.8.
-        aspect_ratio:
-            The ratio of y-unit to x-unit. By default 1.0.
 
 
         TODO
@@ -405,28 +299,9 @@ class Framework(object):
                 "Plotting is currently supported only for 2-dimensional frameworks."
             )
 
-        fig, ax = plt.subplots()
-        ax.set_adjustable("datalim")
-        fig.set_figwidth(canvas_width)
-        fig.set_figheight(canvas_height)
-        ax.set_aspect(aspect_ratio)
-        edge_color_array, edge_list_ref = self.resolve_edge_colors(edge_color)
-
-        nx.draw(
-            self._graph,
-            pos=self.realization(as_points=True, numerical=True),
-            ax=ax,
-            node_size=vertex_size,
-            node_color=vertex_color,
-            node_shape=vertex_shape,
-            with_labels=vertex_labels,
-            width=edge_width,
-            edge_color=edge_color_array,
-            edgelist=edge_list_ref,
-            style=edge_style,
-            **kwargs,
+        self._graph.plot(
+            placement=self.realization(as_points=True, numerical=True), **kwargs
         )
-        plt.show()
 
     @classmethod
     @doc_category("Class methods")
